Create maxBalancedSubstrPartitions.java
We have balanced string str of size N with an equal number of L and R, the task is to find a
maximum number X, such that a given string can be partitioned into X balanced substring. A
string is called to be balanced if the number of ‘L’s in the string equals the number of ‘R’s.